preserve original bbclib.py and bbc_app.py in comapt/ directory for backward compatibility   If you have application code in v 1.1.1 or earlier, mofify to use bbc1.core.compat.bbclib and bbc1.core.compat.bbc_app. codes in bbc1/core/ are modified according to the new design with bbclib_core, bbclib_wire and bbclib.
