+++ tests/compat/test_bbc_app_multi_core_many_transactions_bson.py
@@ -5,11 +5,10 @@ import binascii
 import time
 
 import sys
-sys.path.extend(["../"])
-from bbc1.core import bbclib
+sys.path.extend(["../", "../.."])
+from bbc1.core.compat import bbclib, bbc_app
 from bbc1.core.message_key_types import KeyType
 from bbc1.core.bbc_error import *
-from bbc1.core import bbc_app
 from testutils import prepare, get_core_client, start_core_thread, make_client, domain_setup_utility
