+++ tests/compat/test_bbclib_compat.py
@@ -3,10 +3,10 @@ import pytest
 
 import binascii
 import sys
-sys.path.extend(["../"])
-from bbc1.core.bbclib import BBcTransaction, BBcEvent, BBcReference, BBcWitness, BBcRelation, BBcAsset, \
+sys.path.extend(["../", "../../"])
+from bbc1.core.compat.bbclib import BBcTransaction, BBcEvent, BBcReference, BBcWitness, BBcRelation, BBcAsset, \
     BBcCrossRef, KeyPair, KeyType
-from bbc1.core import bbclib
+from bbc1.core.compat import bbclib
 
 user_id = bbclib.get_new_id("user_id_test1")
 user_id2 = bbclib.get_new_id("user_id_test2")
@@ -26,7 +26,6 @@ event1 = None
 event2 = None
 transaction2 = None
 asset_content = b'abcdefg'
-fmt = bbclib.BBcFormat.FORMAT_BSON_COMPRESS_BZ2
 
 print("\n")
 print("private_key:", binascii.b2a_hex(keypair1.private_key))
@@ -45,15 +44,15 @@ class TestBBcLib(object):
     def test_01_asset(self):
         print("\n-----", sys._getframe().f_code.co_name, "-----")
         global asset1, asset2
-        asset1 = BBcAsset(user_id=user_id, asset_body=b'12345678', format_type=fmt)
-        asset2 = BBcAsset(user_id=user_id, asset_file=asset_content, format_type=fmt)
+        asset1 = BBcAsset(user_id=user_id, asset_body=b'12345678')
+        asset2 = BBcAsset(user_id=user_id, asset_file=asset_content)
 
         # --- for checking serialization function ---
         digest = asset1.digest()
         dat = asset1.serialize()
         print("Digest:", binascii.b2a_hex(digest))
-        print("Serialized data:", dat)
-        asset_tmp = BBcAsset(format_type=fmt)
+        print("Serialized data:", binascii.b2a_hex(dat))
+        asset_tmp = BBcAsset()
         asset_tmp.deserialize(dat)
         print("body_len:", asset_tmp.asset_body_size)
         if asset_tmp.asset_body_size > 0:
@@ -64,15 +63,15 @@ class TestBBcLib(object):
         print("\n-----", sys._getframe().f_code.co_name, "-----")
         print("asset_group_id:", binascii.b2a_hex(asset_group_id))
         global event1, event2
-        event1 = BBcEvent(asset_group_id=asset_group_id, format_type=fmt)
+        event1 = BBcEvent(asset_group_id=asset_group_id)
         event1.add(asset=asset1, mandatory_approver=user_id)
-        event2 = BBcEvent(asset_group_id=asset_group_id, format_type=fmt)
+        event2 = BBcEvent(asset_group_id=asset_group_id)
         event2.add(asset=asset2, mandatory_approver=user_id)
 
         # --- for checking serialization function ---
         dat = event1.serialize()
-        print("Serialized data:", dat)
-        event_tmp = BBcEvent(format_type=fmt)
+        print("Serialized data:", binascii.b2a_hex(dat))
+        event_tmp = BBcEvent()
         event_tmp.deserialize(dat)
         print("mandatory_approvers:", [binascii.b2a_hex(d) for d in event_tmp.mandatory_approvers])
         print("asset_id:", binascii.b2a_hex(event_tmp.asset.asset_id))
@@ -80,13 +79,13 @@ class TestBBcLib(object):
     def test_03_transaction_1(self):
         print("\n-----", sys._getframe().f_code.co_name, "-----")
         global transaction1
-        transaction1 = BBcTransaction(format_type=fmt)
+        transaction1 = BBcTransaction()
         transaction1.add(event=[event1, event2])
-        dummy_cross_ref1 = BBcCrossRef(domain_id=domain_id, transaction_id=transaction1_id, format_type=fmt)
+        dummy_cross_ref1 = BBcCrossRef(domain_id=domain_id, transaction_id=transaction1_id)
         transaction1.add(cross_ref=dummy_cross_ref1)
-        dummy_cross_ref2 = BBcCrossRef(domain_id=domain_id, transaction_id=transaction2_id, format_type=fmt)
+        dummy_cross_ref2 = BBcCrossRef(domain_id=domain_id, transaction_id=transaction2_id)
         transaction1.add(cross_ref=dummy_cross_ref2)
-        witness = BBcWitness(format_type=fmt)
+        witness = BBcWitness()
         transaction1.add(witness=witness)
 
         sig = transaction1.sign(private_key=keypair1.private_key, public_key=keypair1.public_key)
@@ -120,20 +119,19 @@ class TestBBcLib(object):
     def test_04_transaction_with_reference(self):
         print("\n-----", sys._getframe().f_code.co_name, "-----")
         global transaction2, event3, asset3
-        asset3 = BBcAsset(format_type=fmt)
+        asset3 = BBcAsset()
         asset3.add(asset_body=b'bbbbbbb', user_id=user_id)
-        event3 = BBcEvent(asset_group_id=asset_group_id, format_type=fmt)
+        event3 = BBcEvent(asset_group_id=asset_group_id)
         event3.add(asset=asset3, option_approver_num_numerator=1, option_approver_num_denominator=2)
         event3.add(option_approver=user_id)
         event3.add(option_approver=user_id2)
 
-        transaction2 = BBcTransaction(format_type=fmt)
+        transaction2 = BBcTransaction()
         transaction2.add(event=event3)
         reference2 = BBcReference(asset_group_id=asset_group_id,
-                                  transaction=transaction2, ref_transaction=transaction1,
-                                  event_index_in_ref=0, format_type=fmt)
+                                  transaction=transaction2, ref_transaction=transaction1, event_index_in_ref=0)
         transaction2.add(reference=reference2)
-        dummy_cross_ref3 = BBcCrossRef(domain_id=domain_id, transaction_id=transaction1_id, format_type=fmt)
+        dummy_cross_ref3 = BBcCrossRef(domain_id=domain_id, transaction_id=transaction1_id)
         transaction2.add(cross_ref=dummy_cross_ref3)
 
         sig = transaction2.sign(private_key=keypair1.private_key, public_key=keypair1.public_key)
@@ -146,21 +144,20 @@ class TestBBcLib(object):
 
     def test_05_transaction_with_reference2(self):
         print("\n-----", sys._getframe().f_code.co_name, "-----")
-        asset1 = BBcAsset(format_type=fmt)
+        asset1 = BBcAsset()
         asset1.add(user_id=user_id, asset_body=b'ccccc')
-        event = BBcEvent(asset_group_id=asset_group_id, format_type=fmt)
+        event = BBcEvent(asset_group_id=asset_group_id)
         event.add(asset=asset1, option_approver_num_numerator=1, option_approver_num_denominator=2)
         event.add(option_approver=user_id)
         event.add(option_approver=user_id2)
 
         global transaction1
-        transaction1 = BBcTransaction(format_type=fmt)
+        transaction1 = BBcTransaction()
         transaction1.add(event=event)
         reference = BBcReference(asset_group_id=asset_group_id,
-                                 transaction=transaction1, ref_transaction=transaction2,
-                                 event_index_in_ref=0, format_type=fmt)
+                                 transaction=transaction1, ref_transaction=transaction2, event_index_in_ref=0)
         transaction1.add(reference=reference)
-        dummy_cross_ref = BBcCrossRef(domain_id=domain_id, transaction_id=transaction1_id, format_type=fmt)
+        dummy_cross_ref = BBcCrossRef(domain_id=domain_id, transaction_id=transaction1_id)
         transaction2.add(cross_ref=dummy_cross_ref)
 
         sig = transaction1.sign(private_key=keypair2.private_key, public_key=keypair2.public_key)
@@ -178,10 +175,10 @@ class TestBBcLib(object):
 
     def test_06_transaction_with_witness(self):
         print("\n-----", sys._getframe().f_code.co_name, "-----")
-        witness = BBcWitness(format_type=fmt)
+        witness = BBcWitness()
 
         global transaction1
-        transaction1 = BBcTransaction(format_type=fmt)
+        transaction1 = BBcTransaction()
         transaction1.add(witness=witness)
 
         witness.add_witness(user_id)
@@ -203,7 +200,7 @@ class TestBBcLib(object):
 
     def test_06_transaction_with_relation_and_witness(self):
         print("\n-----", sys._getframe().f_code.co_name, "-----")
-        transaction1 = bbclib.make_transaction(relation_num=1, witness=True, format_type=fmt)
+        transaction1 = bbclib.make_transaction(relation_num=1, witness=True)
         bbclib.add_relation_asset(transaction1, relation_idx=0, asset_group_id=asset_group_id,
                                   user_id=user_id, asset_body=b'ccccc')
         bbclib.add_relation_pointer(transaction1, 0, ref_transaction_id=transaction2.digest())
