+++ tests/test_bbc_app_multi_core_compromising_data.py
@@ -8,9 +8,9 @@ import random
 
 import sys
 sys.path.extend(["../"])
-from bbc1.core import bbclib
+from bbc1.core.compat import bbclib
 from bbc1.core.message_key_types import KeyType
-from bbc1.core import bbc_app
+from bbc1.core.compat import bbc_app
 from testutils import prepare, get_core_client, start_core_thread, make_client, domain_setup_utility
