+++ tests/test_bbc_app_multi_core_notification.py
@@ -6,9 +6,8 @@ import time
 
 import sys
 sys.path.extend(["../"])
-from bbc1.core import bbclib
+from bbc1.core import bbclib, bbc_app
 from bbc1.core.message_key_types import KeyType
-from bbc1.core import bbc_app
 from testutils import prepare, get_core_client, start_core_thread, make_client, domain_setup_utility
